+++ src/components/CheckoutForm/styles.css
@@ -93,7 +93,7 @@
 .input-local input {
     margin-bottom: 1rem;
     width: 100%;
-    max-width: 90%;
+    max-width: 100%;
     padding: 0.75rem;
     border-radius: 5px;
     border: 1px solid var(--bege-200);
@@ -260,7 +260,7 @@
     padding: 2rem;
     border-radius: 0 2rem;
     max-width: 400px;
-    max-height: 300px;
+    max-height: 350px;
     width: 100%;
 }
