--- src/pages/Loja/Store/styles.css
@@ -1,304 +1,0 @@
-/* Variáveis de tema exemplo (pode ajustar no seu root ou theme provider) */
-:root {
-    --bege-100: #f8f4ef;
-    --bege-150: #f3e9dd;
-    --bege-200: #f0e0ce;
-    --amarelo-250: #f3c300;
-    --yellow-100: #f1e9c9;
-    --yellow-300: #dbac2c;
-    --yellow-500: #c47f17;
-
-    --white: #fff;
-    --brown-cafe: #6b4e2f;
-    --brown: #3b2f2f;
-    --pretinho: #2d2d2d;
-}
-
-/* Estilos principais */
-.main-container {
-    display: flex;
-    flex-direction: column;
-}
-
-.description-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 2rem;
-    gap: 7rem;
-}
-
-.descricao {
-    max-width: 600px;
-}
-
-.descricao h2 {
-    font-size: 3.2rem;
-    font-weight: 800;
-    line-height: 1.3;
-    color: var(--brown);
-}
-
-.descricao span {
-    font-size: 1.4rem;
-    color: var(--brown-cafe);
-}
-
-.descricao ul {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 1.5rem;
-    margin-top: 2rem;
-    list-style: none;
-    color: var(--brown);
-    padding: 0px;
-}
-
-.descricao li {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    color: var(--gray-500);
-    font-size: 1rem;
-}
-
-.icon-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0.5rem;
-    border-radius: 50%;
-    color: white;
-}
-
-.icon-yellow-300 {
-    background-color: var(--yellow-300);
-}
-
-.icon-yellow-500 {
-    background-color: var(--yellow-500);
-}
-
-.icon-gray-500 {
-    background-color: var(--brown-cafe);
-}
-
-.icon-purple-100 {
-    background-color: var(--brown);
-}
-
-.loja-container {
-    display: flex;
-    flex-direction: column;
-    margin-top: 3rem;
-    font-size: 2rem;
-    color: var(--brown);
-}
-
-.cards {
-    font-size: 1rem;
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 2rem;
-    margin-top: 1rem;
-}
-
-.card {
-    margin-top: 2rem;
-    border-radius: 0 25px;
-    background-color: var(--bege-100);
-    padding: 1rem;
-    width: 15rem;
-    min-height: 19rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    border: 2px solid var(--bege-200);
-}
-
-.card img {
-    margin-top: -60px;
-}
-
-.img-and-tag-local {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.tag-group {
-    display: flex;
-    gap: 0.5rem;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin-top: 0.75rem;
-}
-
-.tag {
-    background-color: var(--bege-200);
-    color: var(--yellow-500);
-    font-weight: bold;
-    border-radius: 100px;
-    padding: 0.25rem 0.75rem;
-    font-size: 0.65rem;
-    text-transform: uppercase;
-}
-
-.description-card-local {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    padding: 1rem 0;
-    gap: 0.5rem;
-    flex: 1;
-}
-
-.title {
-    color: var(--brown);
-    font-weight: bold;
-    font-size: 1.5rem;
-    text-align: center;
-    margin-top: 0.5rem;
-}
-
-.descricao-coffee {
-    color: var(--brown-cafe);
-    text-align: center;
-    margin-top: 0.5rem;
-}
-
-.buy-local {
-    width: 100%;
-    margin-top: auto;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 1rem;
-}
-
-.price {
-    font-size: 1.5rem;
-    color: var(--black);
-}
-
-.price::before {
-    content: "R$ ";
-    font-size: 1rem;
-}
-
-.price strong {
-    font-weight: bold;
-    font-size: 1.5rem;
-    color: var(--gray-600);
-}
-
-.botao-quantidade {
-    background: var(--bege-150);
-    display: flex;
-    align-items: center;
-    border-radius: 6px;
-    padding: 0.5rem;
-}
-
-.botao-quantidade button {
-    background: transparent;
-    border: none;
-    color: var(--yellow-500);
-    font-size: 1.5rem;
-    cursor: pointer;
-    margin: 0 4px;
-}
-
-.botao-quantidade span {
-    margin: 0 0.5rem;
-    color: var(--yellow-500);
-}
-
-.cart-button {
-    background-color: var(--brown);
-    border: none;
-    border-radius: 12px;
-    padding: 0.8rem;
-    cursor: pointer;
-    color: white;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: background-color 0.3s ease, box-shadow 0.3s ease;
-    outline: none;
-    position: relative;
-}
-
-.cart-button:hover {
-    background-color: var(--brown-cafe);
-}
-
-/* Focus apenas se for via teclado */
-.cart-button:focus {
-    box-shadow: 0 0 0 3px var(--brown-cafe);
-}
-
-/* Evita outline forçado ao clicar com mouse */
-.cart-button:focus {
-    outline: none;
-    box-shadow: none;
-}
-
-/* Clique */
-.cart-button:active {
-    transform: scale(0.95);
-}
-
-@media (max-width: 480px) {
-    .descricao h2 {
-        font-size: 3.1rem;
-    }
-
-    .descricao li {
-        font-size: 0.9rem;
-    }
-
-    .card {
-        margin-top: 1rem;
-        width: 17rem;
-    }
-}
-
-@media (max-width: 764px) {
-    .card {
-        margin-top: 1rem;
-        width: 19rem;
-    }
-
-    .loja-container {
-        color: var(--gray-600);
-        text-align: center;
-    }
-}
-
-@media (max-width: 1024px) {
-    .description-container {
-        flex-direction: column;
-        gap: 2rem;
-        text-align: center;
-    }
-
-    .description-container img {
-        max-width: 100%;
-    }
-}
-
-@media (max-width: 1200px), (min-width: 1200px) {
-    .card {
-        margin-top: 1rem;
-        width: 20rem;
-    }
-
-    .loja-container {
-        color: var(--gray-600);
-        text-align: center;
-    }
-}
